ADD Final requirements

+++ src/main/java/com/nttdata/credit/entities/CreditCard.java
@@ -31,9 +31,9 @@ public class CreditCard {
     @Transient
     private Customer customer;
 
-    public void pay(Long quantity){
-        Long amountPaid = this.credit.getAmountPaid();
-        Long amountUsed = this.credit.getAmountUsed();
+    public void pay(Double quantity){
+        Double amountPaid = this.credit.getAmountPaid();
+        Double amountUsed = this.credit.getAmountUsed();
         amountPaid += quantity;
         amountUsed -= quantity;
 
@@ -42,8 +42,8 @@ public class CreditCard {
         this.credit.updateBalance();
     }
 
-    public void consume(Long quantity){
-        Long amountUsed = this.credit.getAmountUsed();
+    public void consume(Double quantity){
+        Double amountUsed = this.credit.getAmountUsed();
         amountUsed += quantity;
         this.credit.setAmountUsed(amountUsed);
         this.credit.updateBalance();
